# src/main/resources/swagger/api.yml
openapi: 3.1.0
info:
  title: Tshepo — Proof-of-Funds Credential API
  version: 0.1.0
  description: |
    Tshepo issues selectively-disclosable SD-JWT Verifiable Credentials attesting
    financial facts computed from a holder's Investec account.  Raw transactions are
    **never persisted** — only boolean claim results are stored.

    ## Two actor surfaces
    | Surface | Auth | Base path |
    |---------|------|-----------|
    | Holder app | JWT (Bearer) | `/api` |
    | Verifier (public) | API key (`X-API-Key`) or none | `/api/verify`, `/v1/verify` |

    ## Issuer identity
    The issuer is identified as a `did:web` DID.  The DID document and JWKS are
    served at the well-known endpoints below and must be resolvable over HTTPS in
    production.

  contact:
    name: Tshepo Engineering
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:8080
    description: Local development (H2 / fixture Investec data)
  - url: https://api.tshepo.app
    description: Production

# ─────────────────────────────────────────────
# Security schemes
# ─────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JHipster-issued JWT returned by `/api/authenticate`.
    apiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: Metered API key for high-volume verifier integrations.

  # ─────────────────────────────────────────────
  # Shared schemas
  # ─────────────────────────────────────────────
  schemas:
...
    BankConnectRequest:
      type: object
      required: [authorizationCode, redirectUri]
      properties:
        authorizationCode:
          type: string
          description: OAuth 2.0 authorization code returned by Investec's auth server.
        redirectUri:
          type: string
          format: uri
          description: Must match the redirect URI registered with Investec.
    BankStatusResponse:
      type: object
      properties:
        connected:
          type: boolean
        connectedAt:
          type: string
          format: date-time
          nullable: true
        maskedAccount:
          type: string
          example: '•••• 4821'
          nullable: true
        status:
          type: string
          enum: [CONNECTED, DISCONNECTED, ERROR]

    # ── Claim types ──────────────────────────────
    ClaimType:
      type: string
      enum: [INFLOW, BALANCE, TENURE, SALARY_CONTINUITY, NO_OVERDRAFT]
...
    ComputedClaim:
      type: object
      properties:
        type:
          $ref: '#/components/schemas/ClaimType'
        operator:
          type: string
        threshold:
          type: number
        currency:
          type: string
        met:
          type: boolean
          description: Whether the holder currently satisfies this threshold.
        basis:
          type: string
          description: Human-readable source description shown to holder only (never in credential).
          example: 'Last 6 months'

    # ── Credential ───────────────────────────────
...
    CredentialResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        holderLogin:
          type: string
        title:
          type: string
        purpose:
          type: string
          nullable: true
        status:
          type: string
          enum: [ACTIVE, REVOKED, EXPIRED]
        issuedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        issuerDid:
          type: string
          example: 'did:web:tshepo.app'
        vcRef:
          type: string
          description: URN reference for the credential.
          example: 'urn:vc:tshepo:8f3a-1d92'
        claims:
          type: array
          items:
            $ref: '#/components/schemas/ComputedClaim'
          description: Claim results shown to holder; never included in presentations.
        claimCount:
          type: integer
    CredentialListResponse:
      type: object
      properties:
        credentials:
          type: array
          items:
            $ref: '#/components/schemas/CredentialResponse'
        total:
          type: integer

    # ── Presentation ─────────────────────────────
...
    VerifierApiKeyResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        label:
          type: string
        active:
          type: boolean
        callCount:
          type: integer
        createdAt:
          type: string
          format: date-time

    # ── DID document ─────────────────────────────
...
    ProblemDetail:
      type: object
      description: RFC 9457 problem detail.
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri

# ─────────────────────────────────────────────
# Paths
# ─────────────────────────────────────────────
paths:
...
  /api/bank/connect:
    post:
      tags: [Bank connection]
      summary: Connect an Investec account
      description: |
        Exchanges the Investec OAuth2 authorization code for an access token,
        pulls account metadata, pre-computes all supported claims in memory,
        then discards the raw transaction data.  Only a `BankConnection` metadata
        record (no transactions) is persisted.
      operationId: connectBank
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BankConnectRequest'
      responses:
        '200':
          description: Bank connected and claims pre-computed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankStatusResponse'
        '400':
          description: Invalid authorization code or redirect URI
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Missing or invalid JWT
        '502':
          description: Investec API unreachable or returned an error
  /api/bank/status:
    get:
      tags: [Bank connection]
      summary: Get current bank connection status
      operationId: getBankStatus
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current connection status for the authenticated holder
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankStatusResponse'
        '401':
          description: Missing or invalid JWT
  /api/bank/claims:
    get:
      tags: [Bank connection]
      summary: List available claims with met/not-met status
      description: |
        Returns all supported claim thresholds with a computed `met` boolean for
        the current holder.  The raw evidence string (`basis`) is included here
        for holder display only — it never enters a credential or presentation.
      operationId: getAvailableClaims
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Available claims
          content:
            application/json:
              schema:
                type: object
                properties:
                  claims:
                    type: array
                    items:
                      $ref: '#/components/schemas/ComputedClaim'
        '401':
          description: Missing or invalid JWT
        '409':
          description: No bank account connected yet

  # ── Credential issuance (holder, JWT-secured) ─
  /api/credentials:
    get:
      tags: [Credentials]
      summary: List holder's credentials
      operationId: listCredentials
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, REVOKED, EXPIRED]
          description: Filter by status.
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated credential list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialListResponse'
        '401':
          description: Missing or invalid JWT
...
  /api/credentials/{id}:
    get:
      tags: [Credentials]
      summary: Get a single credential
      operationId: getCredential
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Credential detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialResponse'
        '401':
          description: Missing or invalid JWT
        '404':
          description: Credential not found or not owned by caller
...
  /api/verifier-keys:
    get:
      tags: [API key management]
      summary: List verifier API keys for the authenticated user
      operationId: listApiKeys
      security:
        - bearerAuth: []
      responses:
        '200':
          description: API key list
          content:
            application/json:
              schema:
                type: object
                properties:
                  keys:
                    type: array
                    items:
                      $ref: '#/components/schemas/VerifierApiKeyResponse'
        '401':
          description: Missing or invalid JWT

    post:
      tags: [API key management]
      summary: Create a new verifier API key
      operationId: createApiKey
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [label]
              properties:
                label:
                  type: string
                  maxLength: 80
      responses:
        '201':
          description: API key created — raw key value returned **once**, store it securely
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/VerifierApiKeyResponse'
                  - type: object
                    properties:
                      rawKey:
                        type: string
                        description: The plaintext API key. Only returned on creation.
        '401':
          description: Missing or invalid JWT
  /api/verifier-keys/{id}:
    delete:
      tags: [API key management]
      summary: Revoke a verifier API key
      operationId: revokeApiKey
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Key revoked
        '401':
          description: Missing or invalid JWT
        '404':
          description: Key not found

# ─────────────────────────────────────────────
# Tags (ordering for docs renderers)
# ─────────────────────────────────────────────
...